requirements.txt
quart
psutil
discord.py
python-dotenv
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
dateparser
requests
